// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/riskAssessmentAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { RiskAssessmentDocumentAC } from './riskAssessmentDocumentAC';
import { RiskAssessmentStatus } from './riskAssessmentStatus';

export interface RiskAssessmentAC extends BaseModelAC { 
    auditableEntity?: string;
    name?: string;
    status: RiskAssessmentStatus;
    statusString?: string;
    year: number;
    summary?: string;
    entityId: string;
    oldId?: string;
    riskAssessmentDocumentACList?: Array<RiskAssessmentDocumentAC>;
    riskAssessmentDocumentFiles?: Array<Blob>;
}