// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/riskAssessmentStatus.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type RiskAssessmentStatus = 0 | 1 | 2 | 3;

export const RiskAssessmentStatus = {
    NUMBER_0: 0 as RiskAssessmentStatus,
    NUMBER_1: 1 as RiskAssessmentStatus,
    NUMBER_2: 2 as RiskAssessmentStatus,
    NUMBER_3: 3 as RiskAssessmentStatus
};